Refactor App.vue
